@@ -5,15 +5,15 @@ authors = [
     {name = "BlueGlassBlock", email = "blueglassblock@outlook.com"},
 ]
 license = {text = "MIT"}
-requires-python = ">=3.8,<4.0"
+requires-python = ">=3.11,<4.0"
 dependencies = [
     "loguru>=0.7.2",
-    "dacite>=1.6.0,<2.0",
+    "dacite>=1.8.0,<2.0",
     "lark[regex]>=1.1,<2.0",
     "typing-extensions<5.0.0,>=4.0.0",
 ]
 name = "kayaku"
-version = "0.5.4"
+version = "0.6.0"
 description = "An powerful and flexible configuration solution."
 readme = "README.md"
 classifiers = [
@@ -23,19 +23,19 @@ classifiers = [
 ]
 
 [tool.pdm]
-includes = ["kayaku"]
+build.includes = ["kayaku"]
 [tool.pdm.scripts]
 test = {composite = ["coverage run -m pytest -vv", "coverage xml", "coverage report"]}
 [tool.pdm.dev-dependencies]
 dev = [
-    "black<23.0.0,>=22.3.0",
-    "isort<6.0.0,>=5.10.1",
+    "black<25.0.0,>=24.10.0",
+    "isort<6.0.0,>=5.13.2",
     "pytest>=7.1.3",
     "coverage>=6.4.4",
-    "rich~=12.5",
+    "rich~=13.9",
     "jsonschema>=4.16.0",
-    "mkdocstrings[python]<1.0.0,>=0.19.0",
-    "mkdocs-material<9.0.0,>=8.4.0",
+    "mkdocstrings[python]<1.0.0,>=0.27.0",
+    "mkdocs-material<10.0.0,>=9.5.44",
     "mkdocs-section-index<1.0.0,>=0.3.4",
     "mkdocs-literate-nav<1.0.0,>=0.4.1",
 ]
