@@ -1,9 +1,4 @@
-from .domain import bootstrap as bootstrap
-from .domain import initialize as initialize
-from .domain import save_all as save_all
-from .model import config as config
-from .model import create as create
-from .model import save as save
+from .manager import Kayaku
 from .utils import copying_field
 
 default = copying_field
