@@ -9,48 +9,61 @@ if base_pth.exists():
 base_pth.mkdir(parents=True, exist_ok=True)
 
 
+def insert_spec(root, src, path) -> None:
+    prefix, suffix = src.prefix, src.suffix
+    target_nd = root.insert(prefix).insert(reversed(suffix))
+    if target_nd.bound:
+        raise ValueError(
+            f"{'.'.join(prefix + ['*'] + suffix)} is already bound to {target_nd.bound}"
+        )
+    target_nd.bound = (src, path)
+
+
 def test_insert_spec():
-    import kayaku.storage
+    import kayaku.bi_tree
     from kayaku.spec import PathSpec, SectionSpec, SourceSpec
 
-    root = kayaku.storage._PrefixNode()
-    kayaku.storage._root.set(root)
+    root = kayaku.bi_tree.Prefix()
     empty = SectionSpec([], [])
 
-    kayaku.storage.insert(
+    insert_spec(
+        root,
         SourceSpec(["a", "b", "c"], ["d", "e", "f"], empty),
         PathSpec([*base_pth.parts, "config"], []),
     )
 
-    kayaku.storage.insert(
+    insert_spec(
+        root,
         SourceSpec(["a", "b", "c"], ["d", "e", "credential"], empty),
         PathSpec([*base_pth.parts, "credential"], []),
     )
 
     with pytest.raises(ValueError):
-        kayaku.storage.insert(
+        insert_spec(
+            root,
             SourceSpec(["a", "b", "c"], ["d", "e", "credential"], empty),
             PathSpec([*base_pth.parts, "credential"], []),
         )
 
 
 def test_lookup_spec():
+    import kayaku.bi_tree
     import kayaku.spec
-    import kayaku.storage
     from kayaku.spec import PathFill, PathSpec, SectionSpec, SourceSpec
 
     empty = SectionSpec([], [])
 
-    root = kayaku.storage._PrefixNode()
-    kayaku.storage._root.set(root)
+    root = kayaku.bi_tree.Prefix()
     path_sect = [PathFill.EXTEND]
 
-    kayaku.storage.insert(
+    insert_spec(
+        root,
         SourceSpec(["a", "b", "c"], ["d", "e", "f"], empty),
         PathSpec([*base_pth.parts, "config"], path_sect),
     )
 
-    kayaku.storage.insert(
+    insert_spec(
+        root,
         SourceSpec(
             ["a", "b", "c", "xxx"],
             [],
@@ -59,17 +72,20 @@ def test_lookup_spec():
         PathSpec([*base_pth.parts, "hmm"], path_sect),
     )
 
-    kayaku.storage.insert(
+    insert_spec(
+        root,
         SourceSpec(["a", "b", "c"], ["d", "e", "credential"], empty),
         PathSpec([*base_pth.parts, "credential"], path_sect),
     )
 
-    kayaku.storage.insert(
+    insert_spec(
+        root,
         SourceSpec(["a", "b", "c", "d"], ["e", "f"], empty),
         PathSpec([*base_pth.parts, "any"], path_sect),
     )
 
-    kayaku.storage.insert(
+    insert_spec(
+        root,
         SourceSpec(["a", "b", "c", "d"], ["o", "p", "e", "f"], empty),
         PathSpec([*base_pth.parts, "whirl"], path_sect),
     )
@@ -100,44 +116,45 @@ def test_lookup_spec():
 
 
 def test_spec_lookup_fmt_err():
-    import kayaku.storage
-    from kayaku.spec import FormattedPath, parse_path, parse_source
+    import kayaku.bi_tree
+    from kayaku.spec import DestWithMount, parse_path, parse_source
 
-    root = kayaku.storage._PrefixNode()
-    kayaku.storage._root.set(root)
-    kayaku.storage.insert(
+    root = kayaku.bi_tree.Prefix()
+    insert_spec(
+        root,
         parse_source("a.b.c.{**}"),
         parse_path(base_pth.as_posix() + "/a/b/c::{}"),
     )
 
-    kayaku.storage.insert(
+    insert_spec(
+        root,
         parse_source("a.b.{**}"),
         parse_path(base_pth.as_posix() + "/d/e/f::{**}"),
     )
 
-    assert (p := root.lookup(["a", "b", "c", "d", "e"])) and p[1] == FormattedPath(
-        Path(base_pth, "d/e/f"), ["c", "d", "e"]
+    assert (p := root.lookup(["a", "b", "c", "d", "e"])) and p[1] == DestWithMount(
+        Path(base_pth, "d/e/f").as_posix(), ("c", "d", "e")
     )
 
 
 def test_spec_lookup_wrapped():
-    import kayaku.storage
-    from kayaku.spec import FormattedPath, parse_path, parse_source
+    import kayaku.bi_tree
+    from kayaku.spec import DestWithMount, parse_path, parse_source
 
-    root = kayaku.storage._PrefixNode()
-    kayaku.storage._root.set(root)
-    kayaku.storage.insert(
+    root = kayaku.bi_tree.Prefix()
+    insert_spec(
+        root,
         parse_source("a.b.c.{**}"),
         parse_path(base_pth.as_posix() + "/a/b/c::{}"),
     )
-    with pytest.raises(ValueError):
-        kayaku.storage.lookup(["a", "b", "c", "d", "e"])
+    assert root.lookup(["a", "b", "c", "d", "e"]) is None
 
-    kayaku.storage.insert(
+    insert_spec(
+        root,
         parse_source("a.b.{**}"),
         parse_path(base_pth.as_posix() + "/d/e/f.jsonc::{**}"),
     )
 
-    assert kayaku.storage.lookup(["a", "b", "c", "d", "e"]) == FormattedPath(
-        Path(base_pth, "d/e/f.jsonc"), ["c", "d", "e"]
+    assert root.lookup(["a", "b", "c", "d", "e"])[1] == DestWithMount(
+        Path(base_pth, "d/e/f.jsonc").as_posix(), ("c", "d", "e")
     )
